+++ NEWS.md
@@ -1,3 +1,5 @@
+# FUNGuildR (development version)
+
 # FUNGuildR 0.2.0
 
 * Update default FUNGuild URL to use new version of FUNguild database (as